leaf57: modularize wasm smoke JS proof
