+++ docs/LEDGER.md
@@ -11,4 +11,4 @@ Allowed status enum: `todo | stubbed | implemented | verified | skipped`.
 | `crates/carreltex-engine/src/tex/tokenize_v0.rs` | engine | tex-tokenizer-v0 | verified | `cargo test --manifest-path crates/carreltex-engine/Cargo.toml` | Deterministic TeX lexing subset with explicit v0 assumptions (NUL invalid, `^^` blocked, `\\verb` blocked, `%` comments, whitespace coalescing, control words/symbols, token cap fail-closed) |
 | `crates/carreltex-wasm-smoke/src/lib.rs` | wasm-adapter | abi-v0 | verified | `./scripts/proof_v0.sh` | Thin ABI adapter over core+engine semantics, strict report/status+missing_components cross-consistency, per-path log bounds + TeX stats JSON invariants with core validator defense-in-depth, deterministic binary events seam carrying kind=1(log bytes)+kind=2(stats JSON), allocator bounded by `MAX_WASM_ALLOC_BYTES_V0` (artifact-aligned), generic artifact-by-name ABI + `main.xdv` copy-out cap enforcement, and mount read-back ABI |
 | `scripts/proof_v0.sh` | proof | v0-bundle | verified | `./scripts/proof_v0.sh` | Bundle gate: LOC guard (scans tracked `crates/**/*.rs` + `scripts/**/*.mjs`, hard limit <=1000 lines) + core tests + wasm smoke + ledger check |
-| `scripts/wasm_smoke_js_proof.mjs` | proof | wasm-js-smoke | verified | `./scripts/proof_wasm_smoke.sh` | JS ABI compatibility checks including compile-request path |
+| `scripts/wasm_smoke_js_proof.mjs` | proof | wasm-js-smoke | verified | `./scripts/proof_wasm_smoke.sh` | Thin JS proof entrypoint; scenarios are modularized under `scripts/wasm_smoke_js/*.mjs` while preserving proof order/output and ABI compatibility checks including compile-request path |
